enable etherscan verification (#12)
* enable etherscan verification

Also add support for custom forge flags via the FORGE_FLAGS env var.

+++ packages/contracts-bedrock/contracts/test/OptimismMintableERC721.t.sol
@@ -134,9 +134,9 @@ contract OptimismMintableERC721_Test is ERC721Bridge_Initializer {
                     "ethereum:",
                     Strings.toHexString(uint160(address(L1Token)), 20),
                     "@",
-                    Strings.toString(1),
+                    Strings.toString(uint256(1)),
                     "/tokenURI?uint256=",
-                    Strings.toString(1)
+                    Strings.toString(uint256(1))
                 )
             )
         );
